Move unit 7 source code into proper directories
All unit 7 practice used to be under practice/exceptions,
but I realized we weren't going to use this for bash/git/github
exercises, so I just got rid of that directory.

I also moved RetrieveElement to be a challenge and updated all of the
moved files' package statements.
