+++ src/com/omegarobotics/unit7/practice/InputMismatch.java
@@ -1,4 +1,4 @@
-package com.omegarobotics.unit7.practice.exceptions;
+package com.omegarobotics.unit7.practice;
 
 /*
  * Prompt the user to enter an integer continuously.
